+++ src/main/java/com/example/music_fly_project/vo/PlayListVO.java
@@ -9,7 +9,7 @@ public class PlayListVO {
     private Long id;
     private long userId;
     private String playListName;
-    private long lengthPlayList;
+    private int lengthPlayList;
     private boolean public1;
     private Date date;
     private byte[] image;
